pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blog-analyzer"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A CLI tool to analyze blog posts and generate manuscripts."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # FastAPI & Server
    "fastapi",
    "uvicorn",
    "python-multipart",

    # Database
    "pymongo",

    # AI SDKs
    "openai",
    "anthropic",
    "google-genai",
    "xai-sdk",

    # AWS
    "boto3",

    # NLP
    "kss",

    # CLI & Utils
    "click",
    "python-dotenv",
    "rich",
    "tqdm",
    "requests",

    # Browser Automation
    "playwright",

    # Image Processing
    "Pillow",
]

[project.optional-dependencies]
gui = [
    "PySide6",
    "streamlit",
]

[project.scripts]
blog-analyzer = "cli:cli"

[tool.setuptools.packages.find]
include = ["llm*", "routers*", "schema*", "utils*", "analyzer*", "ai_lib*", "_prompts*", "_constants*"]
